Pass all tests with solution file
Exercise and readMe are misleading as to the method of "Removing" an
element from an array. This solution creates a whole new array using
a foreach() method on the input array while not including the arguments
presented. The other solution uses the filter method which simplifies
the foreach() solution. If thinking logically to pass the tests one
would filter each individual test and pass all through process of elim.
However, the given solution would MAYBE be reached if looked at from
a different POV. Point is the instructions lead the junior dev to
believe that they would have to filter through all possible tests to
remove the element in the array that was input, not create a whole new
array while omitting the args values.